# lint: align bin/lint with linting in CI