--- a/bin/pre-push
+++ b/bin/pre-push
@@ -16,7 +16,6 @@ set -euo pipefail
 . misc/shlib/shlib.bash
 
 try bin/lint
-try cargo fmt -- --check
 try cargo clippy --all-targets -- -D warnings
 try bin/doc --no-deps