--- a/bin/pre-push
+++ b/bin/pre-push
@@ -15,8 +15,13 @@ set -euo pipefail
 
 . misc/shlib/shlib.bash
 
+# this runs all of lint-fast.sh
 try bin/lint
+
+# part of lint-slow.sh because it requires building code
 try cargo clippy --all-targets -- -D warnings
+
+# part of lint-slow.sh because it requires building code
 try bin/doc --no-deps
 
 try_status_report