--- a/ci/test/lint-main/checks/check-cargo.sh
+++ b/ci/test/lint-main/checks/check-cargo.sh
@@ -19,4 +19,10 @@ cd "$(dirname "$0")/../../../.."
 
 try bin/lint-cargo
 
+try cargo --locked fmt -- --check
+try cargo --locked deny check licenses bans sources
+try cargo hakari generate --diff
+try cargo hakari manage-deps --dry-run
+try cargo deplint Cargo.lock ci/test/lint-deps.toml
+
 try_status_report