--- a/ci/test/lint-main/checks/check-new-line.sh
+++ b/ci/test/lint-main/checks/check-new-line.sh
@@ -8,6 +8,8 @@
 # As of the Change Date specified in that file, in accordance with
 # the Business Source License, use of this software will be governed
 # by the Apache License, Version 2.0.
+#
+# check-new-line.sh — check that files end with an empty line.
 
 set -euo pipefail