--- a/ci/test/lint-main/checks/check-python-files.sh
+++ b/ci/test/lint-main/checks/check-python-files.sh
@@ -9,7 +9,7 @@
 # the Business Source License, use of this software will be governed
 # by the Apache License, Version 2.0.
 #
-# lint — complains about misformatted files and other problems.
+# check-python-files.sh — check Python files for issues.
 
 set -euo pipefail