--- a/ci/test/lint-main/checks/check-rust-version.sh
+++ b/ci/test/lint-main/checks/check-rust-version.sh
@@ -9,7 +9,7 @@
 # the Business Source License, use of this software will be governed
 # by the Apache License, Version 2.0.
 #
-# lint — complains about misformatted files and other problems.
+# check-rust-version.sh — checks the used rust version.
 
 set -euo pipefail