--- a/ci/test/lint-main/checks/check-shell-files.sh
+++ b/ci/test/lint-main/checks/check-shell-files.sh
@@ -8,6 +8,8 @@
 # As of the Change Date specified in that file, in accordance with
 # the Business Source License, use of this software will be governed
 # by the Apache License, Version 2.0.
+#
+# check-shell-files.sh — checks shell files.
 
 set -euo pipefail