This makes sure that `lint-fast.sh` and `bin/lint` do the same. It is a follow-up to https://github.com/MaterializeInc/materialize/pull/23802.

Note that CI additionally runs `lint-slow.sh`, which requires building the code.

### Commits
89574b15c5 lint: complete and update docs of checks
b5a3c3ebf7 lint: extract check-protobuf.sh
5eb28d7079 lint: extract failures in check-crate-license-listing.sh
82e1e08e1c lint: move further cargo checks to check-cargo.sh
8a7dea3b06 lint: improve output
1b96938d01 lint: bin/pre-push: remove command already executed in bin/lint
37501320bc lint: bin/pre-push: add comments
80e5242ccc lint: extract check-python-formatting.sh from check-python-formatting.sh
d43d0da3d2 lint: order checks by name

### Overview of checks after the changes
```
ci/test/lint-main
├── after
│   └── check-no-diff.sh
├── before
│   └── can-lint.sh
└── checks
    ├── check-cargo.sh
    ├── check-copyright.sh
    ├── check-crate-license-listing.sh
    ├── check-generation.sh
    ├── check-new-line.sh
    ├── check-protobuf.sh
    ├── check-python-docs.sh
    ├── check-python-files.sh
    ├── check-python-formatting.sh
    ├── check-rust-test-attributes.sh
    ├── check-rust-version.sh
    └── check-shell-files.sh
```

### Performance impact

Performance of `bin/lint` deteriorated from 9 seconds to 14 seconds compared to the introduction of parallel linting with https://github.com/MaterializeInc/materialize/pull/23802. Linting is still faster than it was before the introduction of parallel linting (26 seconds).

The slowest steps are `check-python-files.sh`, `check-crate-license-listing.sh`, and `check-new-line.sh`, which take over 12 seconds. Interestingly, `check-python-files.sh` and `check-new-line.sh` are slower a bit since this PR, possibly due to much concurrent IO access.

#### Before
9 seconds (multiple measurements conducted).

```
--- Running checks in /Users/nrainer/Workspaces/mz-repo/ci/test/lint-main/before
--- can-lint.sh (SUCCEEDED in 0.03s)
--- Running checks in /Users/nrainer/Workspaces/mz-repo/ci/test/lint-main/checks
--- check-python-docs.sh (SUCCEEDED in 3.40s)
--- check-shell-files.sh (SUCCEEDED in 2.68s)
--- check-copyright.sh (SUCCEEDED in 6.34s)
--- check-python-files.sh (SUCCEEDED in 10.16s)
--- check-rust-files.sh (SUCCEEDED in 5.01s)
--- check-cargo.sh (SUCCEEDED in 0.21s)
--- check-rust-version.sh (SUCCEEDED in 0.03s)
--- check-new-line.sh (SUCCEEDED in 7.97s)
--- check-generation.sh (SUCCEEDED in 0.94s)
--- Running checks in /Users/nrainer/Workspaces/mz-repo/ci/test/lint-main/after
--- check-no-diff.sh (SUCCEEDED in 0.06s)
+++ Linting result
All checks successful.
```

#### After
14 seconds (multiple measurements conducted).

```
--- Running 1 check(s) in ci/test/lint-main/before
--- can-lint.sh (SUCCEEDED in 0.03s)
--- Running 11 check(s) in ci/test/lint-main/checks
--- check-python-docs.sh (SUCCEEDED in 4.65s)
--- check-shell-files.sh (SUCCEEDED in 2.95s)
--- check-protobuf.sh (SUCCEEDED in 4.32s)
--- check-copyright.sh (SUCCEEDED in 11.62s)
--- check-python-files.sh (SUCCEEDED in 13.84s)
--- check-cargo.sh (SUCCEEDED in 6.54s)
--- check-crate-license-listing.sh (SUCCEEDED in 12.64s)
--- check-rust-version.sh (SUCCEEDED in 0.03s)
--- check-new-line.sh (SUCCEEDED in 13.30s)
--- check-rust-test-attributes.sh (SUCCEEDED in 7.02s)
--- check-generation.sh (SUCCEEDED in 1.18s)
--- Running 1 check(s) in ci/test/lint-main/after
--- check-no-diff.sh (SUCCEEDED in 0.07s)
+++ Linting result
All checks successful.
```

#### Other

I further extracted `check-python-formatting.sh` from `check-python-formatting.sh`:
```
--- check-python-files.sh (SUCCEEDED in 10.65s)
--- check-python-formatting.sh (SUCCEEDED in 5.39s)
```

However, the overall duration remains 14 seconds due to step `check-crate-license-listing.sh`, which cannot be split.