@nrainer-materialize 2023-12-15
@philip-stoev: Could you do me a favor by checking out this branch and trying to run `bin/lint`? If I recall correctly, it did not work immediately for me but required installing some cargo stuff. If that is the case, I would like to document that and possibly allow disabling checks with system vars.