@philip-stoev 2023-12-15
@nrainer-materialize , yes it did not work for me out of the box:

```
pstoev@Ubuntu-2004-focal-64-minimal:~/nrainer-materialize-lint/align-local-linting-with-ci$ ./ci/test/lint-main/checks/check-crate-license-listing.sh
--- cargo --locked about generate ci/deploy/licenses.hbs
error: no such command: `about`

	Did you mean `logout`?

	View all installed commands with `cargo --list`
+++ Status report
0/1 commands passed
```


```
--- git branch --show-current
0
1
--- Fetch target branch
$ git fetch origin main
From github.com:nrainer-materialize/materialize
 * branch                  main       -> FETCH_HEAD
--- Verify that protobuf config is up-to-date
--- bin/pyactivate ./ci/test/lint-buf/generate-buf-config.py
Working dir: /home/pstoev/nrainer-materialize-lint/align-local-linting-with-ci
Collected 72 proto files.
15 proto files to be ignored from breaking check.
Written buf configuration to './src/buf.yaml'.
--- yamllint src/buf.yaml
--- git diff --name-only --exit-code src/buf.yaml
--- Lint protobuf
$ git merge-base HEAD origin/main
--- buf breaking src --against .git#ref=5b16629d73071098b4e495f37e52b560080de4c9,subdir=src --verbose
misc/shlib/shlib.bash: line 78: buf: command not found
^^^ 🚨 Failed: buf breaking src --against .git#ref=5b16629d73071098b4e495f37e52b560080de4c9,subdir=src --verbose
+++ Status report
3/4 commands passed
```

```
pstoev@Ubuntu-2004-focal-64-minimal:~/nrainer-materialize-lint/align-local-linting-with-ci$ ./ci/test/lint-main/checks/check-cargo.sh
--- bin/lint-cargo
--- cargo --locked fmt -- --check
--- cargo --locked deny check licenses bans sources
bans ok, licenses ok, sources ok
--- cargo hakari generate --diff
error: no such command: `hakari`

	View all installed commands with `cargo --list`
^^^ 🚨 Failed: cargo hakari generate --diff
--- cargo hakari manage-deps --dry-run
error: no such command: `hakari`

	View all installed commands with `cargo --list`
^^^ 🚨 Failed: cargo hakari manage-deps --dry-run
--- cargo deplint Cargo.lock ci/test/lint-deps.toml
error: no such command: `deplint`

	View all installed commands with `cargo --list`
^^^ 🚨 Failed: cargo deplint Cargo.lock ci/test/lint-deps.toml
+++ Status report
3/6 commands passed
```

I think that would be ok, meaning my dev environment was not properly setup. The problem is that none of the error outputs above appeared when running `bin/lint`. `bin/lint` just reported a list of failures but then I had to run the bash scripts manually to obtain the actual error output.