@philip-stoev 2023-12-15
> I would like to document that 

Here is what I was missing:

```
cargo install cargo-hakari --locked
cargo install cargo-about --locked
cargo install cargo-deplint --locked
```

This leaves `buf`, the installation of which is OS-specific.