@nrainer-materialize 2023-12-15
Asked for feedback in https://materializeinc.slack.com/archives/CM7ATT65S/p1702651476994149.