@nrainer-materialize 2023-12-18
Added:
3c0278087e lint: check whether tooling for linting is installed
1e295b6eee guide: list required tooling for linting